Implemented AutoEventWireup basically
There is now a base page class that should be included on the page.  It
expects two events "Page_Init" and "Page_Load" to be declared on the
page.  If it doesn't find them, then nothing happens. There is now
rudimentary support for wiring up button onclick events.  The name of
the events should be buttonID_Click.  That is the ID of the button
assigned when calling New_Button, NOT the name of the variable.
(Although in most cases this should be the same thing)